Minor perf tweaks on LeadingZeroCount
