+++ src/HdrHistogram/HistogramBase.cs
@@ -875,7 +875,7 @@ namespace HdrHistogram
         }
         private static int GetBucketIndex(long value, long subBucketMask, int bucketIndexOffset)
         {
-            var leadingZeros = NumberOfLeadingZeros(value | subBucketMask); // smallest power of 2 containing value
+            var leadingZeros = Bitwise.NumberOfLeadingZeros(value | subBucketMask); // smallest power of 2 containing value
             return bucketIndexOffset - leadingZeros;
         }
 
@@ -896,25 +896,6 @@ namespace HdrHistogram
             return ValueFromIndex(bucketIndex, subBucketIndex);
         }
 
-        private static int NumberOfLeadingZeros(long value)
-        {
-            // Code from http://stackoverflow.com/questions/9543410/i-dont-think-numberofleadingzeroslong-i-in-long-java-is-based-floorlog2x/9543537#9543537
-
-            // HD, Figure 5-6
-            if (value == 0)
-                return 64;
-            var n = 1;
-            // >>> in Java is a "unsigned bit shift", to do the same in C# we use >> (but it HAS to be an unsigned int)
-            var x = (uint)(value >> 32);
-            if (x == 0) { n += 32; x = (uint)value; }
-            if (x >> 16 == 0) { n += 16; x <<= 16; }
-            if (x >> 24 == 0) { n += 8; x <<= 8; }
-            if (x >> 28 == 0) { n += 4; x <<= 4; }
-            if (x >> 30 == 0) { n += 2; x <<= 2; }
-            n -= (int)(x >> 31);
-            return n;
-        }
-        
         private IRecordedData GetData()
         {
             var relevantCounts = GetRelevantCounts();
